fix(ui consistency): Improved UI consistency in file metadata display and fixed type conve...
# Project State

## Current Focus
Improved UI consistency in file metadata display and fixed type conversion in Git TUI section rendering

## Context
The changes standardize the file metadata display format in the file manager and fix a type conversion issue in the Git TUI's section rendering.

## Completed
- [x] Refactored file metadata display in file manager to use aligned columns for labels and values
- [x] Fixed type conversion in Git TUI section rendering from usize to u16

## In Progress
- [ ] No active work in progress

## Blockers
- None identified

## Next Steps
1. Verify UI consistency across all file manager views
2. Test Git TUI section rendering with various file counts
---
category: fix
scope: ui consistency
---

+++ examples/git_tui.rs
@@ -599,7 +599,7 @@ impl GitTui {
 
             let mut row = sub_y;
             if !staged.is_empty() {
-                render_section_card(plane, 2, row, 40, staged.len() + 3, t);
+                render_section_card(plane, 2, row, 40, (staged.len() + 3) as u16, t);
                 draw_text(
                     plane,
                     3,
@@ -638,7 +638,7 @@ impl GitTui {
             }
 
             if !modified.is_empty() {
-                render_section_card(plane, 2, row, 40, modified.len() + 3, t);
+                render_section_card(plane, 2, row, 40, (modified.len() + 3) as u16, t);
                 draw_text(
                     plane,
                     3,
